db_upgrade.py
#!flask/bin/python

from migrate.versioning import api
from config import config

api.upgrade(config.SQLALCHEMY_DATABASE_URI, config.SQLALCHEMY_MIGRATE_REPO)
v = api.db_version(config.SQLALCHEMY_DATABASE_URI, config.SQLALCHEMY_MIGRATE_REPO)

print('Current database version: ' + str(v))