requirements.txt
alembic==0.9.1
aniso8601==1.2.1
blinker==1.4
click==6.7
Flask==0.12.1
Flask-DebugToolbar==0.10.1
flask-marshmallow==0.7.0
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
marshmallow==2.13.5
marshmallow-sqlalchemy==0.13.1
psycopg2==2.7.1
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.12.1
